requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiosqlite==0.20.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
stripe==9.5.0
python-dotenv==1.0.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
email-validator==2.1.1
